# food-retail-marketplace/ardo-backend/docs/swagger.yaml
definitions:
...
  core.MlString:
    additionalProperties:
      type: string
    type: object
...
  management_product.AddProductIn:
    properties:
      desc:
        $ref: '#/definitions/core.MlString'
      img:
        type: string
      name:
        $ref: '#/definitions/core.MlString'
      price:
        type: number
      quantity:
        type: integer
      status:
        type: string
    type: object
  management_product.Product:
    properties:
      createdAt:
        type: string
      desc:
        $ref: '#/definitions/core.MlString'
      id:
        type: string
      img:
        type: string
      isDeleted:
        type: boolean
      name:
        $ref: '#/definitions/core.MlString'
      price:
        type: number
      quantity:
        type: integer
      status:
        type: string
      updatedAt:
        type: string
      version:
        type: integer
    type: object
  management_product.ProductsData:
    properties:
      count:
        type: integer
      products:
        items:
          $ref: '#/definitions/management_product.Product'
        type: array
    type: object
  management_product.UpdateProductIn:
    properties:
      desc:
        $ref: '#/definitions/core.MlString'
      img:
        type: string
      name:
        $ref: '#/definitions/core.MlString'
      price:
        type: number
      quantity:
        type: integer
      status:
        type: string
    type: object
  management_section.CreateSectionIn:
    properties:
      img:
        type: string
      name:
        $ref: '#/definitions/core.MlString'
    type: object
  management_section.Section:
    properties:
      createdAt:
        type: string
      id:
        type: string
      img:
        type: string
      isDeleted:
        type: boolean
      name:
        $ref: '#/definitions/core.MlString'
      updatedAt:
        type: string
      version:
        type: integer
    type: object
  management_section.SectionsData:
    properties:
      count:
        type: integer
      sections:
        items:
          $ref: '#/definitions/management_section.Section'
        type: array
    type: object
  management_section.UpdateSectionIn:
    properties:
      img:
        type: string
      name:
        $ref: '#/definitions/core.MlString'
    type: object
...
  response.Error:
    properties:
      messages:
        example:
        - error message
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
      traceId:
        example: string
        type: string
    type: object
  response.Success:
    properties:
      data: {}
      success:
        example: true
        type: boolean
      traceId:
        example: string
        type: string
    type: object
host: api.ardogroup.org
info:
  contact:
    name: API Support
    url: https://nassyrov.t.me
  description: Detailed info about all endpoints
  title: Ardo Backend API
  version: 1.0.0
paths:
...
  /management/products:
    get:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Is deleted
        in: query
        name: is_deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/management_product.ProductsData'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get products by filters
      tags:
      - Management Products
    post:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: New product object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/management_product.AddProductIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Add product
      tags:
      - Management Products
  /management/products/{product_id}:
    delete:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Delete product
      tags:
      - Management Products
    get:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/management_product.Product'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get product by id
      tags:
      - Management Products
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Update product object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/management_product.UpdateProductIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update product
      tags:
      - Management Products
  /management/products/recover/{product_id}:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Recover product
      tags:
      - Management Products
...
  /management/sections:
    get:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Is deleted
        in: query
        name: is_deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/management_section.SectionsData'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get sections by filters
      tags:
      - Management Sections
    post:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Create section object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/management_section.CreateSectionIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Create section
      tags:
      - Management Sections
  /management/sections/{section_id}:
    delete:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Delete section
      tags:
      - Management Sections
    get:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Success'
            - properties:
                data:
                  $ref: '#/definitions/management_section.Section'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get section by id
      tags:
      - Management Sections
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      parameters:
      - description: Update section object
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/management_section.UpdateSectionIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Update section
      tags:
      - Management Sections
  /management/sections/recover/{section_id}:
    put:
      consumes:
      - application/json
      description: This can only be done by the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        default:
          description: ""
          schema:
            $ref: '#/definitions/response.Error'
      summary: Recover section
      tags:
      - Management Sections
...
schemes:
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"